Split up model and controller
