+++ src/main/java/gui/phylogeny/NewickNodeMouseEventHandler.java
@@ -1,5 +1,6 @@
 package gui.phylogeny;
 
+import gui.phylogeny.model.NewickNode;
 import javafx.event.Event;
 import javafx.event.EventHandler;
 import javafx.scene.input.MouseEvent;
