+++ src/main/java/gui/phylogeny/model/NewickColourMatching.java
@@ -1,4 +1,4 @@
-package gui.phylogeny;
+package gui.phylogeny.model;
 
 import javafx.scene.paint.Paint;
